@MasonProtter 2020-03-25
@wongalvis14 and I discussed on Slack. Here are timings on my machine for 
* CuArrays v1.7.2
```julia
(@v1.4) pkg> add CuArrays#v1.7.2
   Updating registry at `~/.julia/registries/General`
   Updating git-repo `https://github.com/JuliaRegistries/General.git`
  Resolving package versions...
   Updating `~/.julia/environments/v1.4/Project.toml`
  [3a865a2d] + CuArrays v1.7.2 #v1.7.2 (https://github.com/JuliaGPU/CuArrays.jl.git)
   Updating `~/.julia/environments/v1.4/Manifest.toml`
  [3895d2a7] + CUDAapi v3.1.0
  [c5f51814] + CUDAdrv v6.0.0
  [be33ccc6] + CUDAnative v2.10.2
  [3a865a2d] + CuArrays v1.7.2 #v1.7.2 (https://github.com/JuliaGPU/CuArrays.jl.git)
  [0c68f7d7] + GPUArrays v2.0.1
  [929cbde3] + LLVM v1.3.4
  [a759f4b9] + TimerOutputs v0.5.3julia> using BenchmarkTools, CuArrays
[ Info: Precompiling CuArrays [3a865a2d-5b23-5a0f-bc46-62713ec82fae]
┌ Warning: Incompatibility detected between CUDA and LLVM 8.0+; disabling debug info emission for CUDA kernels
└ @ CUDAnative ~/.julia/packages/CUDAnative/hfulr/src/CUDAnative.jl:114
WARNING: using CuArrays.BLAS in module Main conflicts with an existing identifier.julia> using BenchmarkTools, CuArraysjulia> function pi_mc_cu(nsamples)
           xs = CuArrays.rand(nsamples); ys = CuArrays.rand(nsamples)
           mapreduce((x, y) -> (x^2 + y^2) < 1.0, +, xs, ys, init=0) * 4/nsamples
       end
pi_mc_cu (generic function with 1 method)julia> @benchmark pi_mc_cu(10000000)
[ Info: Building the CUDAnative run-time library for your sm_75 device, this might take a while...
BenchmarkTools.Trial: 
  memory estimate:  4.61 KiB
  allocs estimate:  126
  --------------
  minimum time:     594.163 μs (0.00% GC)
  median time:      658.573 μs (0.00% GC)
  mean time:        671.493 μs (2.87% GC)
  maximum time:     2.311 ms (55.14% GC)
  --------------
  samples:          7424
  evals/sample:     1
```
* CuArrays Master (today)
```julia
(@v1.4) pkg> add https://github.com/JuliaGPU/CuArrays.jl.git#master
   Updating git-repo `https://github.com/JuliaGPU/CuArrays.jl.git`
  Resolving package versions...
   Updating `~/.julia/environments/v1.4/Project.toml`
  [3a865a2d] + CuArrays v2.0.0 #master (https://github.com/JuliaGPU/CuArrays.jl.git)
   Updating `~/.julia/environments/v1.4/Manifest.toml`
  [3895d2a7] + CUDAapi v4.0.0
  [c5f51814] + CUDAdrv v6.2.1
  [be33ccc6] + CUDAnative v3.0.1
  [f68482b8] + Cthulhu v1.0.0
  [3a865a2d] + CuArrays v2.0.0 #master (https://github.com/JuliaGPU/CuArrays.jl.git)
  [0c68f7d7] + GPUArrays v3.1.0
  [929cbde3] + LLVM v1.3.4
  [dc548174] + TerminalMenus v0.1.0
  [a759f4b9] + TimerOutputs v0.5.3julia> using BenchmarkTools, CuArrays
[ Info: Precompiling CuArrays [3a865a2d-5b23-5a0f-bc46-62713ec82fae]
WARNING: using CuArrays.BLAS in module Main conflicts with an existing identifier.julia> function pi_mc_cu(nsamples)
           xs = CuArrays.rand(nsamples); ys = CuArrays.rand(nsamples)
           mapreduce((x, y) -> (x^2 + y^2) < 1.0, +, xs, ys, init=0) * 4/nsamples
       end
pi_mc_cu (generic function with 1 method)julia> @benchmark pi_mc_cu(10000000)
[ Info: Building the CUDAnative run-time library for your sm_75 device, this might take a while...
BenchmarkTools.Trial: 
  memory estimate:  7.81 KiB
  allocs estimate:  245
  --------------
  minimum time:     10.014 ms (0.00% GC)
  median time:      10.159 ms (0.00% GC)
  mean time:        10.198 ms (0.31% GC)
  maximum time:     11.559 ms (9.85% GC)
  --------------
  samples:          491
  evals/sample:     1
```
* Alvis's PR branch:
```julia
(@v1.4) pkg> add https://github.com/wongalvis14/CuArrays.jl.git#mapreduce
   Updating git-repo `https://github.com/wongalvis14/CuArrays.jl.git`
   Updating registry at `~/.julia/registries/General`
   Updating git-repo `https://github.com/JuliaRegistries/General.git`
  Resolving package versions...
   Updating `~/.julia/environments/v1.4/Project.toml`
  [3a865a2d] + CuArrays v2.0.0 #mapreduce (https://github.com/wongalvis14/CuArrays.jl.git)
   Updating `~/.julia/environments/v1.4/Manifest.toml`
  [3895d2a7] + CUDAapi v4.0.0
  [c5f51814] + CUDAdrv v6.2.1
  [be33ccc6] + CUDAnative v3.0.1
  [f68482b8] + Cthulhu v1.0.0
  [3a865a2d] + CuArrays v2.0.0 #mapreduce (https://github.com/wongalvis14/CuArrays.jl.git)
  [0c68f7d7] + GPUArrays v3.1.0
  [929cbde3] + LLVM v1.3.4
  [dc548174] + TerminalMenus v0.1.0
  [a759f4b9] + TimerOutputs v0.5.3julia> using BenchmarkTools, CuArrays
[ Info: Precompiling CuArrays [3a865a2d-5b23-5a0f-bc46-62713ec82fae]
WARNING: using CuArrays.BLAS in module Main conflicts with an existing identifier.julia> function pi_mc_cu(nsamples)
           xs = CuArrays.rand(nsamples); ys = CuArrays.rand(nsamples)
           mapreduce((x, y) -> (x^2 + y^2) < 1.0, +, xs, ys, init=0) * 4/nsamples
       end
pi_mc_cu (generic function with 1 method)julia> @benchmark pi_mc_cu(10000000)
BenchmarkTools.Trial: 
  memory estimate:  11.58 KiB
  allocs estimate:  357
  --------------
  minimum time:     7.527 ms (0.00% GC)
  median time:      7.715 ms (0.00% GC)
  mean time:        7.795 ms (0.52% GC)
  maximum time:     10.703 ms (13.98% GC)
  --------------
  samples:          642
  evals/sample:     1
```