@MasonProtter 2020-03-26
@maleadt mentioned in #611 that it could be because `mapreduce(f, op, xs, ys)` is falling back on `redude(op, map(f, xs, ys))` but looking at the old implementation: https://github.com/JuliaGPU/CuArrays.jl/blob/v1.7.2/src/mapreduce.jl I don't see any obvious mechanism in the old version to fuse the map with reduce when there's multiple containers.