@wongalvis14 2020-03-27
This implementation is faster than the old one on 1D array mapreduce

v1.7
```
julia> @benchmark pi_mc_cu(100000000)
BenchmarkTools.Trial: 
  memory estimate:  3.75 KiB
  allocs estimate:  113
  --------------
  minimum time:     4.068 ms (0.00% GC)
  median time:      7.378 ms (0.00% GC)
  mean time:        25.011 ms (0.15% GC)
  maximum time:     698.870 ms (0.00% GC)
  --------------
  samples:          200
  evals/sample:     1
```

New impl:
```
julia> @benchmark pi_mc_cu(100000000)
BenchmarkTools.Trial: 
  memory estimate:  14.91 KiB
  allocs estimate:  426
  --------------
  minimum time:     2.941 ms (0.00% GC)
  median time:      5.904 ms (0.00% GC)
  mean time:        20.800 ms (0.47% GC)
  maximum time:     723.428 ms (0.00% GC)
  --------------
  samples:          253
  evals/sample:     1
```